6th commit Major refactoring – update.
The following new types and interfaces added into the IFormField.ts file: dynControlType, standardInputType, IFormFieldValidator, IFormOptions. The IFormField interface updated with new properties. They will also be used in the future with additional components and functionality Also, some properties have the type defined in the added types and interfaces. The ICategory interface has been added dataObjects subfolder.

The ‘DataService’ updated with the getCategories() method for obtaining an array of ICategory objects from the backend.

The ‘ItemFormFieldsService’ has a major update:
The ‘$item’ Signal removed and the new ‘$formFields’ Signal has been added.
The set signal method is being used to update new signal values.
For updating the $formSignal (array of objects) we have 2 options: we can either using de-structuring using the spread operator ‘…’ when we update the formFields array, or we can use a custom equal comparison function (e.g.: the defined function ‘equalPrimitives’). Otherwise, the $formFields signal does not send a change notification to the FormComponent.
The getCategories() method of the ‘DataService’ is called in the constructor, and then sets the initial option values of the corresponding ‘select’ field.
The setItemId() updated, to call the ‘updateFormFieldsInitialValues’ method.
The following new domain-logic methods have been added:
‘setFormFieldSelectOptions’ (gets from backend an array of objects and updates the initial option values of a specified ‘select’ field - then sets the updated formField array to the $formFields signal),
‘updateFormFieldsInitialValues’ (updates the formFields with initial values from the IItem object fetched, and then sets a new value to the $formFields signal - then sets the updated formField array to the $formFields signal),
‘updateSelectOptions’ (updates/sets as true the option values of the specified ‘select’ field on the basis of the values fetched with the IItem object),
‘updateRadioOptions’ (respectively, updates/sets as true the value of the specified ‘radio’ field on the basis of the value fetched with the IItem object),
‘dateTimeString’ (Just convers the Date value from the IItem object to an ISO date-time string for setting the initial value of a date-time picker field).

The ‘FormComponent’ has been refactored to accept a ‘’ready-made” ‘formFields’ array as a Signal from the ‘ItemFormFieldsService’ instead of an ‘IItem’ object. It monitors signal changes into the effect() method in the constructor. Whenever the $formFields signal is changed, the setFormControlValues() is used to patch the form controls’ values accordingly.

Comments are left for clarity. They will be removed in next commits.