11th commit The abstract class ‘FormFieldsService’ (form-fields-service.ts) renamed to ‘BaseFormFieldsService’ (base-form-fields.service.ts) The ‘FormFieldsModule’ (form-fields.module) has been deleted/removed.
The new ‘FormFieldsFacroryService’ has been created into the ‘service’ subfolder.
It is known only to interested in components: the ‘RequestDataComponent’ -the initiator- and the ‘FormComponent’ -the consumer.
The ‘FormFieldsFacroryService’ takes an injection token into its constructor, and on the basis of the token’s value, it instantiates (just once) the appropriate form fields domain service (ffService).
The ‘FormFieldsFacroryService’ exposes a public method, the setId() method, so if a consumer wants to use this method, the ‘FormFieldsFacroryService’ should have been previously instantiated. This method targets an ‘initiator’ consumer, and it actually passes the id value to the appropriate domain service instantiated, for formFields array preparation.
The ‘FormFieldsFacroryService’ also exposes a public static method, the getFormFieldsService(), that returns the “ready” instantiated domain service. This method targets the real consumer, e.g. the FormComponent(s), which then can obtain the formFields array from that domain service.

The ‘RequestDataComponent’ (‘the initiator’) is getting access to the ‘FormFieldsFacroryService’, by adding it, as well as the preferred token and its value, to the providers array section of the @Component. Doing so, the ‘FormFieldsFacroryService’ is instantiated (by Angular DI) and can be injected to the ‘RequestDataComponent’ component. Actually, whenever the RequestDataComponent’ component is activated, the ‘FormFieldsFacroryService’ is alos instantiated. The ‘FormFieldsFacroryService’ instantiation, also means that the appropriate domain service is also getting instantiated. Then, the ‘RequestDataComponent’ assigns the instance of the specific domain service (in the ngOnInit life hook method) and uses it, to pass the id, every time it changes.

The ’FormComponent’ can get the instance of the instantiated domain service in its constructor (via the static method ‘FormFieldsFacroryService.getFormFieldsService()’), and then it can use it (as previously), to access the formFields array via domain service signal (‘$formFields’), and continuously listen to changes, by registering an effect() function. Note that the ’FormComponent’ does not instantiate the ’FormComponent’.  The ’FormComponent’ should have been instantiated by the ‘initiator’ -the RequestDataComponent’. If it is NOT, then a button ‘Error !!!’ is displayed as a ‘form’.