14th commit A new toolbar with a couple of mat-button links have been added into the HomeComponent template. The router outlet has been also added to the HomeComponent template, and the code file has been accordingly updated as well. A bit of changes also made in the stylesheet.
The routes array in the app.router.ts file have been updated. The dtToken has been added as static data to the items and categories routes.The new component AboutComponent for matching the added About button/link, but it is just for demo purposes and might be updated later on. The empty path ‘’ for the HomeComponent route remains unused.

The code class and the template of the AppComponent have been updated, and aligned to the updated router routes/paths.

The RequestDataComponent has been updated. The main change is that, now in the constructor, it subscribes to the activated router to get the passed-in data. In our case this is just the dtToken passed as static data in the routes array. Once the dtToken is updated then the appropriate domain service is instantiated via the FormFieldsFacroryServiceProvider.
Subscribing to the constructor is necessary because we get an instance of a service (a domain service like the ItemFormFieldsService or the CategoryFormFieldsService ), which is actually a service that is injected into the RequestDataComponent.